feat(media): added Episode entity for provide video

+++ MediaService/src/main/java/me/elpomoika/MediaService/dto/media/RatingRequest.java
@@ -1,4 +1,4 @@
 package me.elpomoika.MediaService.dto.media;
 
-public record RatingRequest(Double rating) {
+public record RatingRequest(int rating) {
 }
